move folder auth to pages folder and add max-len rule , and fix styling for signup page relates #5 ;
