+++ .eslintrc.json
@@ -15,5 +15,6 @@
         "@typescript-eslint"
     ],
     "rules": {
+        "max-len": ["error", { "code": 80 }]
     }
 }
