+++ client/.eslintrc.json
@@ -26,6 +26,7 @@
     "rules": {
         "no-nested-ternary": 0,
         "no-use-before-define": "off",
+        "max-len": ["error", { "code": 80 }],
         "@typescript-eslint/no-use-before-define": ["error"],
         "react/jsx-filename-extension": [ "warn", {"extensions": [".tsx"]} ],
         "func-names": ["error", "never"],
